Split oxidation headers
